Fix a crash when showing stash while patch review plugin is not loaded.
The stash manager (in the Git Plugin) lists available stashed patches. When
a user clicked on a stash item, the stash manager tried to use the patch
review plugin to show the stash item. However, it did not check that it
was actually loaded and this lead to a crash when the patch review plugin
was not enabled.

This commit tests whether the plugin is available. If not, it opens the
stashed patch as a plain diff document.