SourceFormatterStyle::copyDataFrom: fix documentation
fe268683f2221495750ce015207d66527c213f56 added copying of m_usePreview
in this function, but failed to update the documentation. List the data
that is not copied by this function instead of what is copied, because
it should be clear why name, caption and description are not copied to
a derived style.