Refactor ItemRepository to explicitly set mutex from outside
Instead of having m_ownMutex and a way to optionally override
that from the outside, always force the mutex to be provided
from the outside.

This is by itself not a functional change, but it will allow us
to get rid of the (deprecated) recursive mutices one by one, just
like we have established already in the indexedstring.cpp code.
I.e. all of the mutices we are providing are still recursive.