Declare FileModificationCache as Q_MOVABLE_TYPE
0c3d13a89d14a8d6dfca8ec6e1434f3b1f6b7395 removed this declaration
because in Qt 5 it is useful only for types stored in QVector,
QVarLengthArray and QList (and not QHash where FileModificationCache is
a mapped_type). However, QHash has been reimplemented in Qt 6 with some
optimizations enabled when both the key and the value type are
relocatable. Declaring FileModificationCache as Q_MOVABLE_TYPE makes no
difference now, but will benefit performance once KDevelop is ported to
Qt 6.