Distinguish between read/write ops to item repository
Basically replace the single ::op call with two variants:

1. read, taking a `const& repo`
2. write, taking a `& repo`

This clarifies the intention and makes the code more resilient.
While doing this change, I reviewed the usages and fix a few
cases where we took a mutable repo but only want/need a const ref.

Thanks to Igor Kushnir for the suggestion!