Parse qmake `else {}` statements
Add a basic test to verify that this succeeds now.

This is the first step in unbreaking the test_qmakeproject now that
e.g. gcc-base.conf uses this syntax. We did not support it at all
previously, leading to bogus parse errors.

No semantic analysis is added here, so the existing project parse
test still fails.