Unbreak test_qmakeproject
The test was... completely broken? I really have no clue how it
could have ever worked. It tried to lookup the target item by its
path, but targets have no path. And definitely not a path in the
build dir?

Furthermore, for subdirs we don't have an "all" target, we only
have the individual targets for dir_a and dir_b...

Adapt the test such that it works. Quite frankly noone should be
using QMake anyways, and I spend far too much time on QMake here.
Let's just make it pass and be done with it for now...