Skip cmake configure when importing projects with up2date project data
When we open a cmake project, try to import the file data directly.
If that works and the project data was generated after the last
modification of any cmake files that are used by the project, use
the data as-is. This way, we don't need to wait for CMake configure
when opening a project in KDevelop in many cases, greatly improving
the perceived performance.

When we manually reload an item, we still force cmake to run. This
is mostly a safety-net for now. I don't think performance matters
a lot in this case, as we either need to reconfigure anyways because
we are reacting to a file change event, or because the user explicitly
triggered a reload which might indicate something was broken.