Configurable semantic colors
Add a page under _Color Themes -> Highlighting Text Styles -> KDevelop/Semantic Colors_
that allows the configuration of the semantic highlight colors.
Additionally, add a "global color source" setting to the language
settings. When the latter is set to "From Editor Theme" we will
use the colors from the theme as-is without any further blending.
_Bold font for declarations_ will still make the declarations bold
in addition to the bold settings from the config dialog.

Local rainbox colorization is not affected by this at all.

See also:
https://invent.kde.org/kdevelop/kdevelop/uploads/500467bc355906d363e3a76a9044e138/ColorConfig.png

BUG: 395856