Revert "Configurable semantic colors"
This negatively influences the default look and feel of the semantic
colors. I'm all for offering users the ability to fine tune the looks,
but it's not acceptable that the default is so bad that everyone would
have to edit the colors manually.

We'll have to come up with a solution to this problem before we can
merge this back in again.

This reverts commit 54b3be95cd66f9cf0141031a5d97d4b6a5dc8031.