Unify template arguments for ItemRepositoryFor
The type we was is just a context, it doesn't have to be the item.
Unify on the non-item class for which the repository is used, or
introduce a separate context type tag where none exists. This allows
for shorter callee code.

Note that the separate type tag is already an established paradigm
which was used previously for the DeclarationComment e.g.

Fixes: https://invent.kde.org/kdevelop/kdevelop/-/issues/9