Don't clear the persistent symbol table cache early
DUChainPrivate::doMoreCleanup() clears the cache before the destruction
of objects with static object duration. The crash described in
a8ffba2c470f0ecd91f15c6ac5ced6315e8a3d64 indicates that the cache is
filled again afterwards. So the only effect of clearing the cache seems
to be slower exit because of more cache misses.