Handle operators when extracting template parameters
Complex identifiers might include fancy expressions including
operator functions. In such cases we got totally confused by
operators that use angle brackets. This patch fixes that by
detecting such cases and skipping them.

In theory we should only do that for C++, in practice this is somewhat
hard to detect at this level. Furthermore, I believe we shouldn't run
into such issues with other languages anyways, so handling it
generically is hopefully fine.

This patch also fixes runtime warnings of the form:
```
Using QCharRef with an index pointing outside the valid range of a QString.
The corresponding behavior is deprecated, and will be changed in a future version of Qt.
```