Optimize: prevent double isOperator lookups in findClose
Skip ahead by one char and only look at the remaining text, instead
of handling the current char twice, once in findCommaOrEnd and then
directly again in findClose when called from there.

Doing so requires we properly handle strings in a different way,
which is easy to do. While at it, we also fix this for chars,
which did not get handled correctly in findCommaOrEnd before.

This was suggested by Igor Kushnir during code review, many thanks!