Fix another situation where we could go past-end in ParamIterator
When we set m_curEnd to the end but m_cur is left at a previous stage,
we would end up dereferencing past the end of the string in operator++

Fixes:
```
Using QCharRef with an index pointing outside the valid range of a QString.
The corresponding behavior is deprecated, and will be changed in a future version of Qt.
```