Set KDEV_PLUGIN_VERSION to KDEVELOP_SOVERSION
5d3cdf9f39a3916ae9753b3d78633bd65f3e815f bumped KDevelop_VERSION_MINOR
to 10, which increased KDEVELOP_SOVERSION from 59 to 510 and
KDEV_PLUGIN_VERSION from 30+59-52=37 to 30+510-52=488. A plugin version
jump to 510 is not much worse than to 488. The advantage is an obvious
correspondence between the two versions instead of an obscure and now
arbitrary relation between them.

Hopefully decreasing KDEVELOP_SOVERSION and KDEV_PLUGIN_VERSION to 60
when KDevelop 6.0 is released won't be a problem.