Code completion: Fix '.' to '->' for some cases
There are some situations where KDevelop erroneously replaces a '.' with
a '->', e.g. for floating point numbers some specific aggregate
initializers.
This fixes this (at least for some situations) by checking if clang
reports a parse issue with a dot-to-arrow substitution.
A new unit test is added to check the correct behaviour for the
situation described in bug 460870. To check this, the code
completion unit tests needed to be extended a bit: It is now possible
to pass the expected code to the test that is then compared with the
actual result.

BUG: 460870
FIXED-IN: 5.11.230400