Don't static_cast QSortFilterProxyModel* to TreeModel*
3 slots of AsyncTreeView static_cast its model() to TreeModel*. But in
fact, model() is a QSortFilterProxyModel or a VariableSortProxyModel.
Neither of the two actual types of model() inherits TreeModel. This
undefined behavior of a wrong casting was introduced by setting a proxy
as the model in fa956301ecf0de5dc533f1c90cf21372476297c8. I have noticed
only one consequence of this undefined behavior in practice: when an
item is expanded, TreeModel::itemChildrenReady is not emitted, and thus
tree columns are not resized - neither in a variable tooltip nor in
Variables tool view.

Changes in this commit:
* Store a reference to the TreeModel in AsyncTreeView to get rid of the
  casting.
* A new virtual function AsyncTreeView::mapViewIndexToTreeModelIndex()
  replaces AsyncTreeView's dependence on QSortFilterProxyModel, which
  allows to remove a useless proxy model from VariableToolTip. The proxy
  model is useless because it never does any sorting or filtering. It
  was introduced to satisfy AsyncTreeView's interface.
* Replace unused data member VariableTree::m_model with a more useful
  member function AsyncTreeView::treeModel().