Fix misleading comment about refilling the grep job history
In commit b9904391 performing the past grep/search jobs again at
session start was changed to simply restoring the settings that
are needed to rerun a grep job. A comment in the constructor
of the grep tool view was not changed accordingly and may lead to
the false assumption that the search is automatically performed again.