Make Variables Toolview column's auto-resize configurable
Add a toolbar to the Variables Toolview to let user configure if
they want columns to be automatically resized when they click in
the Variables tree.

The class DebuggerToolFactoryBase serves the purpose to factor
code that is not dependent on the template parameter of its
subclasses, therefore avoid re-instantiating the same code for
each different template instantiation.

CCBUG: 388658