cmake: introduce a cache of files with no compilation data
Opening Configure Project dialog or analyzing a file with Cppcheck while
a CMake project is the current project or analyzing a CMake project with
Cppcheck, produces many identical debug messages in KDevelop's output:
    kdevelop.plugins.cmake: no information found for ""
That's because the cppcheck plugin collects include directories for each
project target, and a target item's path() is empty. This is a bug in
the cppcheck plugin, which should be fixed, but it could remain unfixed
for a long time. This change optimizes such fruitless queries for
empty-path items, and the debug message is printed once now.

The following debug message is now printed once rather than 6 times for
each .cpp file without compilation data in a project with zero targets:
    kdevelop.plugins.cmake: no information found for "/path/to/file"

The added mutable data member should not break thread safety, because
there is no locking when the data is modified. Apparently
CMakeManager::fileInformation() is always called from the main thread.
The definitions of DefinesAndIncludesManager's member functions that
(indirectly) call fileInformation() start with the following assertion:
    Q_ASSERT(QThread::currentThread() == qApp->thread());