cmake: show KJob::errorString(), not errorText() in the UI
From the documentation for KJob::errorText():
  This is usually some extra data associated with the error, such as a
  URL. Use errorString() to get a human-readable, translated message.

The definitions of KJob::errorText() and KJob::errorString() are
identical:
    return d_func()->errorText;
But errorString() is virtual and should be implemented by derived
classes. Currently only one job in KDevelop's source code overrides
errorString(): QMakeJob. QMakeJob is never a subjob of
ChooseCMakeInterfaceJob, therefore there is no practical consequence of
this commit for now.