cmake: show a warning if project is configured with outdated data
When project data is outdated right after configuring, KDevelop does not
reconfigure the project to avoid a reload loop in case of a bug or some
upstream CMake change. Project data can be outdated right after
configuring if the user edits a source CMake file during a CMake
configure or generate step. In this case reloading the project will
bring the data up to date. Show a warning that asks the user to do just
that. The user can finish editing source CMake files, reload the
project, and the reminding warning will disappear automatically. If a
reload loop occurs for some reason, the user will eventually cease
reloading and hopefully report a bug.

I plan to implement the added TODO together with future project
reloading fixes.

CCBUG: 460450