nativeDialog's parent: OpenProjectDialog => main window
With OpenProjectDialog as the parent, the native Open Project file
dialog always appears in the top left corner of the leftmost monitor.
With OpenProjectDialog's parent (the main window) as the parent, this
native dialog is centered on top of KDevelop's main window.

Note that subsequent assistant dialog pages were and are centered on top
of KDevelop's main window. An assistant dialog page OpenProjectPage is
shown instead of the native file dialog in KDE Plasma. That page is
centered on top of KDevelop's main window too.

Quote from QDialog's documentation:
Note that QDialog (and any other widget that has type Qt::Dialog) uses
the parent widget slightly differently from other classes in Qt. A
dialog is always a top-level widget, but if it has a parent, its default
location is centered on top of the parent's top-level widget (if it is
not top-level itself). It will also share the parent's taskbar entry.

The native Open Project file dialog should be destroyed along with its
logical owner OpenProjectDialog. As OpenProjectDialog is no longer the
parent, a ScopedDialog must be stored instead of the raw pointer.
Creating the native file dialog in OpenProjectDialog's member
initializer list would be awkward => implement null ScopedDialog
constructor and ScopedDialog::assign().